newco/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

South>=0.7.4
psycopg2==2.4.5

django-social-auth>=0.6.6
oauth2>=1.5.211
selenium>=2.20.0
django-shop-configurableproduct>=0.3.1
